services/resume-service/requirements.txt
fastapi
uvicorn[standard]

# --- Supabase Client (Storage and DB) --- 
supabase

# ---  Pydantic for Settings and Data Models ---
pydantic
pydantic-settings

# ---  AI & Scraping Services ---
google-genai
landingai
agentic-doc         
firecrawl-py     
# ---  Utilities ---
python-dotenv
python-multipart      

# --- Development & Testing Dependencies ---
pytest              
httpx
pytest-asyncio              
